requirements.txt
# By using Conda
#conda-forge::fastmcp
#conda-forge::openai
#conda-forge::pyautogui
#conda-forge::pywinauto

# Or pip
mss==10.1.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comtypes
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
distro 
fastcrc==0.3.4
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
geographiclib==2.1
geopy==2.4.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
Jinja2==3.1.6
jiter
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mcp==1.22.0
mdurl==0.1.2
MouseInfo
mpmath==1.3.0
networkx==3.6
numpy==2.2.6
openai
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
polars==1.35.2
polars-runtime-32==1.35.2
psutil==7.1.3
PyAutoGUI
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyGetWindow
Pygments==2.19.2
PyJWT==2.10.1
pymavlink==2.4.49
PyMsgBox
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
pyparsing==3.2.5
pyperclip
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.1
PyRect
PyScreeze
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
python-xlib
pytweening
pywin32==311
pywinauto
PyYAML==6.0.3
pyzbar==0.1.9
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.29.0
rubicon-objc
scikit-fuzzy==0.5.0
scipy==1.16.3
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.3
starlette==0.50.0
sympy==1.14.0
torch==2.9.1
torchvision==0.24.1
tqdm
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.3.233
ultralytics-thop==2.0.18
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
